feat(clients): Add API for listing all attached clients.
This adds a new endpoint, `/account/attached_clients`, which lists all
clients attached to the user's account - devices, sessions, and oauth -
in a single call. It uses state that's not currently exposed via the API
in order to merge the list in a sensible way, e.g. by ensuring that each
OAuth refresh token shows up as a separate entry, but that a device record
with corresponding refresh token does not get listed twice.

The new endpoint is intended to replace the existing `/sessions` and
`/client-tokens` endpoints, which are currently used by the front-end
for generating the "devices and apps" list.